MechaControl 1.0
Moves Forward/Backward/Left/Right with the arrows, and rotates with q and d (bc I work on an AZERTY keyboard)
Though, the mecha (it actually looks more like a capsule but still) will follow the same directions no matter how much it turned around, which can be a little disturbing when you don't know it. I'll fix it later. For now, I gotta EAT.