Enemy AI fixed and now consider line of sight
Now the Enemy AI will stop following the player if it is too close AND can see the player. However, it doesn't consider line of sight for other kinds of targets, like simple positions. NOTE: The AI will have a strange behavior if both the enemy and the mech are standing in the opposite sides of a really, really thin wall. But the wall has to be so thin that's it's irrevelant. Also, some constants will have to adjusted depending on both player's and enemy's size.